+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+# version 0.0.5
+
+- fixed bug that made rremind print the archive folder at every call
+
+- added "version" command
+
+- added syntax "rremind -3..8" 
+
 # version 0.0.4
 
-- implemented rudimental editing of config file
+- implemented rudimentary editing of config file
 
 - implemented archiving
 
@@ -26,9 +34,9 @@
 
 # Unreleased
 
-[ ] Implement "version" command.
 [ ] Make "archive" accept a parameter specifying how old the appointments need to be for archiving.
 [ ] Make archive configurable so that it collects all archives in *one* file.
+[ ] Make archive *report* what it's doing
 [ ] "sort" command to sort the rem files chronologically.
 [ ] "put" command to add appointments.
 [ ] Accept date-range input: rremind 0..7 to show appointments of the next 7 days.
@@ -38,6 +46,7 @@
 
 # History
 
+[x] Implement "version" command. [v 0.0.5]
 [x] Make path to .rem-folder mutable [v 0.0.4]
 [x] Bug: "When" fails to give the time of the appointment [v 0.0.4]
 [x] archive dates that are past [v 0.0.4]